requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4
pandas==2.3.2
xlrd>=2.0.1
numpy==2.3.3
google-generativeai==0.8.5
databricks-sql-connector==4.1.3
fastapi==0.117.1